Creating new module for core-kotlin-strings (2)
