--- core-kotlin-modules/core-kotlin-strings/src/main/kotlin/com/baeldung/numeric_strings/NumericStrings.kt
@@ -1,16 +1,0 @@
-package com.baeldung.numeric_strings
-
-object NumericStrings {
-    fun isNumericToX(toCheck: String): Boolean {
-        return toCheck.toDoubleOrNull() != null
-    }
-
-    fun isNumericRegex(toCheck: String): Boolean {
-        val regex = "-?[0-9]+(\\.[0-9]+)?".toRegex()
-        return toCheck.matches(regex)
-    }
-
-    fun isNumericAllAndIsDigit(toCheck: String): Boolean {
-        return toCheck.all { char -> char.isDigit() }
-    }
-}
